Changed Node-Red network mode to host (#215)
Node red does not work well when it is on bridge mode with other tools
such as HomeBridge and HomeAssistant. Running it on host mode solves
problem. The other alternative is to have both on same network but this
one is better since homebridge asks host mode specifically which is one
of the main integration which people use with Node-Red.