Add ``Virtual Machine Manager`` to App Store (May Need Changes on Host) (#467)
This PR adds Virtual Machine Manager (virt-manager) which uses libvirt
and QEMU/KVM to run virtual machines. It depends on having QEMU and
libvirt on the host though, but this can easily be configured also to
run it from the container itself by rebuilding it with QEMU and libvirt
included. However, having those installed on the host is better for the
best performance in the VMs that are being managed from the container.
Also, it resolves conflicts. By making sure that it runs from the host,
someone can also connect as client to QEMU from their own system (option
to connect to remote libvirt server with QEMU over SSH on host). If we'd
do this in the container, we need it to have a seperate IP from the host
and it also needs to include SSH. This is not ideal, so it's better to
have QEMU and libvirt on the host, mount the files for them to work in
the container and connect to it. This keeps the GUI possibility of this
container while also allowing the client to connect from their local
Virtual Machine Manager app directly. The Docker image has the
``latest`` tag though, but it's the only pushed tag. So I don't think
this will cause conflicts in this case. It works for me from that tag.
Is there a way we can also still hookup a script into this app to
install QEMU and libvirt on the host? Or like add notes?

---------

Co-authored-by: CorrectRoad <a778917369@gmail.com>