Add playit Agent (#547)
This adds the agent of playit to the CasaOS App Store. The function of
this container and app would be to setup an agent that can connect to
playit.gg's reverse proxies to from then setup what ports to tunnel from
local to playit.gg (and use their provided subdomain to access it
outside of the network where the server is hosted). Check the before
install tips for instructions on how to use it. This has no screenshots
included since it is a non-GUI container.